move capture code closer to equivalent PL code
