+++ pennylane/capture/primitives.py
@@ -17,17 +17,16 @@ created in pennylane.
 
 It has a jax dependency and should be located in a standard import path.
 """
+from pennylane._grad import _get_grad_prim, _get_jacobian_prim
 from pennylane.control_flow.for_loop import _get_for_loop_qfunc_prim
 from pennylane.control_flow.while_loop import _get_while_loop_qfunc_prim
+from pennylane.measurements.capture_measurements import _get_abstract_measurement
 from pennylane.measurements.mid_measure import _create_mid_measure_primitive
+from pennylane.operation import _get_abstract_operator
 from pennylane.ops.op_math.adjoint import _get_adjoint_qfunc_prim
 from pennylane.ops.op_math.condition import _get_cond_qfunc_prim
 from pennylane.ops.op_math.controlled import _get_ctrl_qfunc_prim
 from pennylane.workflow._capture_qnode import qnode_prim
-
-from .capture_diff import _get_grad_prim, _get_jacobian_prim
-from .capture_measurements import _get_abstract_measurement
-from .capture_operators import _get_abstract_operator
 
 AbstractOperator = _get_abstract_operator()
 AbstractMeasurement = _get_abstract_measurement()
